[!!!][FEATURE] Improved cache handling
The handling of the permission cache is improved. The caching framework
is used to store the calculated user permissions. This also allows
the usage of different cache backends (like redis).

The cache handling is centralized so that the cache is cleared for
all uses when a permission is edited or a new page is added.

!!! Requires at least TYPO3 6.0.